Move getName to translationUtils
Refs #397.
