+++ app/components/reservation/ReservationsListItem.js
@@ -4,9 +4,9 @@ import { Link } from 'react-router';
 
 import TimeRange from 'components/common/TimeRange';
 import ReservationControls from 'containers/ReservationControls';
-import { renderReservationStateLabel } from 'utils/renderUtils';
-import { getName } from 'utils/DataUtils';
 import { getCaption, getMainImage } from 'utils/imageUtils';
+import { renderReservationStateLabel } from 'utils/renderUtils';
+import { getName } from 'utils/translationUtils';
 
 class ReservationsListItem extends Component {
   renderImage(image) {
