+++ app/screens/shared/resource-list/ResourceListItem.js
@@ -2,8 +2,9 @@ import React, { Component, PropTypes } from 'react';
 import Glyphicon from 'react-bootstrap/lib/Glyphicon';
 import { Link } from 'react-router';
 
-import { getHumanizedPeriod, getName } from 'utils/DataUtils';
+import { getHumanizedPeriod } from 'utils/DataUtils';
 import { getMainImage } from 'utils/imageUtils';
+import { getName } from 'utils/translationUtils';
 import ReserveButton from './ReserveButton';
 
 class ResourceListItem extends Component {
