+++ app/utils/translationUtils.js
@@ -1,3 +1,7 @@
+function getName(item, language) {
+  return getProperty(item, 'name', language);
+}
+
 function getProperty(item, property, language = 'fi') {
   if (item && item[property] && item[property][language]) {
     return item[property][language];
@@ -6,5 +10,6 @@ function getProperty(item, property, language = 'fi') {
 }
 
 export {
+  getName,
   getProperty,
 };
